% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_ofratiooftwomeans.R
\name{CI_of_ratio_oftwo_means_droprows}
\alias{CI_of_ratio_oftwo_means_droprows}
\title{Wald Ratio Calculator With Confidence Intervals}
\usage{
CI_of_ratio_oftwo_means_droprows(
  data,
  mean_colA,
  mean_colB,
  se_colA,
  se_colB,
  t
)
}
\arguments{
\item{data}{A dataframe}

\item{mean_colA}{Mean of the variable as numerator}

\item{mean_colB}{Mean of the variable as denominator}

\item{se_colA}{Standard error of the variable as numerator}

\item{se_colB}{Standard error of the variable as denominator}

\item{t}{t statistic for the desired confidence}
}
\value{
A dataframe with additonal columns
}
\description{
Wald Ratio Calculator With Confidence Intervals
}
